feat: Move DevTools Selector Parser to  CompletionEngine
